Update Header class references and add new project references
Refactored the code to replace usage of the 'Header' class with the 'CubeDatagramHeader' class in MainWindow.xaml.cs for better readability and maintainability. This change involves modifying the reading and writing methods of the aforementioned class. Additionally, included new package references in the LedCube.Streamer.csproj file. This was needed to implement logging and hosting features. Also added a project reference to 'LedCube.Core.csproj'. Moreover, an unnecessary folder reference in LedCube.Core.UI.csproj was removed.